Fix issue where the number of logged crafts was lower than the actual number of crafts
This changeset adds more accurate counting of the number of crafts that
have been done. It was possible for one craft to be counted multiple times,
and for multiple crafts to be counted as one craft depending on the timing
of TRADE_SKILL_ITEM_CRAFTED_RESULT events and the player's framerate.

For me this was easiest to observe when crafting potions or flasks by
starting a run of 10 crafts, then observing the "Crafts" number on the
advanced craft log showing something lower than 10.

This fix in turn makes the statistics related to crafting costs and profit
more accurate.