Ensure professionStats data is correct after doing AverageProfit calculations
This changeset fixes an issue where the professionStats on recipeData
are modified in-place to calculate the impact of resourcefulness/multicraft
points in the AverageProfit module. The changes were reverted, but the
old values were still used in later calculations because recipeData:Update()
wasn't called.

This can be observed on salvage recipes when using the bountiful phials that
add resourcefulness points during winter months. The bountiful phial bonus
resourcefulness is not recognized without this change.